Update RCV projects new api
Chapter 3-4 with new api
