+++ RVisualTutorial/TutorialFunnelPlot/chapter3_RCustomVisual/funnelRvisual_v01/tsconfig.json
@@ -8,7 +8,7 @@
     "out": "./.tmp/build/visual.js"
   },
   "files": [
-    ".api/v1.7.0/PowerBI-visuals.d.ts",
+    ".api/v1.10.0/PowerBI-visuals.d.ts",
     "node_modules/powerbi-visuals-utils-dataviewutils/lib/index.d.ts",
     "src/settings.ts",
     "src/visual.ts"
