+++ RVisualTutorial/TutorialFunnelPlot/chapter3_RCustomVisual/funnelRvisual_v03/capabilities_old.json
@@ -1,0 +1,98 @@
+{
+  "dataRoles": [
+    {
+      "displayName": "Population",
+      "description": "Numeric variable (positive integer)",
+      "kind": "GroupingOrMeasure",
+      "name": "population"
+    },
+    {
+      "displayName": "Occurrence",
+      "description": "Numeric variable (nonnegative integer). Must be smaller than 'Population'",
+      "kind": "GroupingOrMeasure",
+      "name": "occurrence"
+    },
+    {
+      "displayName": "Tooltips",
+      "description": "Fields to be used as mouseover tooltips",
+      "kind": "Grouping",
+      "name": "tooltips"
+    }
+  ],
+   "dataViewMappings": [
+    {
+      "conditions": [
+        {
+          "population": {
+            "max": 1
+          },
+          "occurrence": {
+            "max": 1
+          },
+          "tooltips": {
+            "max": 7
+          }
+        }
+      ],
+      "scriptResult": {
+        "dataInput": {
+          "table": {
+            "rows": {
+              "select": [
+                {
+                  "for": {
+                    "in": "population"
+                  }
+                },
+                {
+                  "for": {
+                    "in": "occurrence"
+                  }
+                },
+                {
+                  "for": {
+                    "in": "tooltips"
+                  }
+                }
+              ],
+              "dataReductionAlgorithm": {
+                "top": {}
+              }
+            }
+          }
+        },
+        "script": {
+          "scriptProviderDefault": "R",
+          "scriptOutputType": "png",
+          "source": {
+            "objectName": "rcv_script",
+            "propertyName": "source"
+          },
+          "provider": {
+            "objectName": "rcv_script",
+            "propertyName": "provider"
+          }
+        }
+      }
+    }
+  ],
+  "objects": {
+    "rcv_script": {
+      "properties": {
+        "provider": {
+          "type": {
+            "text": true
+          }
+        },
+        "source": {
+          "type": {
+            "scripting": {
+              "source": true
+            }
+          }
+        }
+      }
+    }
+  },
+  "suppressDefaultTitle": true
+}
