+++ RVisualTutorial/TutorialFunnelPlot/chapter3_RCustomVisual/funnelRvisual_v03/pbiviz.json
@@ -2,14 +2,14 @@
   "visual": {
     "name": "funnelRvisualVer03",
     "displayName": "funnelRvisualVer03",
-    "guid": "funnelRvisualVer03",
+    "guid": "funnelRvisualVer03_FEFF3",
     "visualClassName": "Visual",
-    "version": "1.0.3",
+    "version": "1.0.0",
     "description": "",
     "supportUrl": "",
     "gitHubUrl": ""
   },
-  "apiVersion": "1.7.0",
+  "apiVersion": "1.10.0",
   "author": {
     "name": "",
     "email": ""
