+++ RVisualTutorial/TutorialFunnelPlot/chapter3_RCustomVisual/funnelRvisual_v03/script_old.r
@@ -242,7 +242,20 @@ if(validToPlot && !gpd) # too small canvas
   sizeWarn = 8 #smaller 
 }
 
- 
+
+#RVIZ_IN_PBI_GUIDE:BEGIN:Removed to enable custom visual fields 
+# if(validToPlot && (ncol(dataset)< 2)) # invalid input 
+# {
+#   validToPlot = FALSE
+#   pbiWarning1 = "Both population and occurrence are required"
+#   pbiWarning = cutStr2Show(pbiWarning1, strCex = sizeWarn/6, partAvailable = 0.9)
+# }
+
+# if(validToPlot)
+#   validData = complete.cases(dataset) & (dataset[,1]>=dataset[,2])
+#RVIZ_IN_PBI_GUIDE:END:Removed to enable custom visual fields
+
+#RVIZ_IN_PBI_GUIDE:BEGIN:Added to enable custom visual fields 
 if(validToPlot && (!exists("population") ||!exists("occurrence"))) # invalid input 
 {
   validToPlot = FALSE
@@ -271,6 +284,7 @@ if(validToPlot && (sum(validData) < minPoints)) # not enough data samples
   pbiWarning2 = cutStr2Show(pbiWarning2, strCex = sizeWarn/6, partAvailable = 0.9)
   pbiWarning<-paste(pbiWarning1, "<br>", pbiWarning2, sep="")
 }
+#RVIZ_IN_PBI_GUIDE:END:Added to enable custom visual fields 
 
 
 
@@ -289,13 +303,15 @@ if(validToPlot && (sum(validData) < minPoints)) # not enough data samples
 
 if(validToPlot)
 {
+  #RVIZ_IN_PBI_GUIDE:BEGIN:Added to enable custom visual fields   
   if(!exists("tooltips"))
   {
     dataset = cbind(population,occurrence)
   }else{
     dataset = cbind(population,occurrence,tooltips)
   }
-
+  #RVIZ_IN_PBI_GUIDE:END:Added to enable custom visual fields
+  
   dataset = dataset[validData,]# keep only valid
   namesDS = names(dataset)
