+++ RVisualTutorial/TutorialFunnelPlot/chapter4_RHTMLCustomVisual/funnelRHTMLvisual_v01/.vscode/settings.json
@@ -19,19 +19,19 @@
       "fileMatch": [
         "/pbiviz.json"
       ],
-      "url": "./.api/v1.7.0/schema.pbiviz.json"
+      "url": "./.api/v1.10.0/schema.pbiviz.json"
     },
     {
       "fileMatch": [
         "/capabilities.json"
       ],
-      "url": "./.api/v1.7.0/schema.capabilities.json"
+      "url": "./.api/v1.10.0/schema.capabilities.json"
     },
     {
       "fileMatch": [
         "/dependencies.json"
       ],
-      "url": "./.api/v1.6.0/schema.dependencies.json"
+      "url": "./.api/v1.10.0/schema.dependencies.json"
     }
   ]
 }
